add "/" and "/[platform]"
